+++ src/pymetropolis/metro_demand/modes/car/__init__.py
@@ -2,16 +2,12 @@ from .car_driver import CarDriverPreferencesStep
 from .car_driver_with_passengers import CarDriverWithPassengersPreferencesStep
 from .car_passenger import CarPassengerPreferencesStep
 from .car_ridesharing import CarRidesharingPreferencesStep
-from .distances import CarFreeFlowDistancesStep, CarShortestDistancesStep
 from .files import (
     CarDriverPreferencesFile,
     CarDriverWithPassengersPreferencesFile,
-    CarFreeFlowDistancesFile,
     CarFuelFile,
-    CarODsFile,
     CarPassengerPreferencesFile,
     CarRidesharingPreferencesFile,
-    CarShortestDistancesFile,
 )
 from .fuel_consumption import CarFuelStep
 
@@ -20,9 +16,6 @@ CAR_FILES = [
     CarDriverWithPassengersPreferencesFile,
     CarPassengerPreferencesFile,
     CarRidesharingPreferencesFile,
-    CarODsFile,
-    CarShortestDistancesFile,
-    CarFreeFlowDistancesFile,
     CarFuelFile,
 ]
 
@@ -31,7 +24,5 @@ CAR_STEPS = [
     CarDriverWithPassengersPreferencesStep,
     CarPassengerPreferencesStep,
     CarRidesharingPreferencesStep,
-    CarShortestDistancesStep,
-    CarFreeFlowDistancesStep,
     CarFuelStep,
 ]