+++ src/pymetropolis/metro_demand/modes/car/fuel_consumption.py
@@ -1,9 +1,10 @@
 import polars as pl
 
+from pymetropolis.metro_demand.routing.files import TripsCarFreeFlowTravelTimesFile
 from pymetropolis.metro_pipeline import Step
 from pymetropolis.metro_pipeline.parameters import FloatParameter
 
-from .files import CarFreeFlowDistancesFile, CarFuelFile
+from .files import CarFuelFile
 
 
 class CarFuelStep(Step):
@@ -15,15 +16,17 @@ class CarFuelStep(Step):
         "fuel.consumption_factor", description="Fuel consumption, in liters per km."
     )
     fuel_price = FloatParameter("fuel.price", description="Price of fuel, in € per liter.")
-    input_files = {"distances": CarFreeFlowDistancesFile}
+    input_files = {"ff_distances": TripsCarFreeFlowTravelTimesFile}
     output_files = {"fuel_consumption": CarFuelFile}
 
     def is_defined(self) -> bool:
         return self.fuel_factor is not None
 
     def run(self):
-        df: pl.DataFrame = self.input["distances"].read()
-        df = df.select("trip_id", fuel_consumption=self.fuel_factor * pl.col("distance") / 1000.0)
+        df: pl.DataFrame = self.input["ff_distances"].read()
+        df = df.select(
+            "trip_id", fuel_consumption=self.fuel_factor * pl.col("free_flow_distance") / 1000.0
+        )
         # If `fuel_price` is not defined (None), `fuel_cost` will be all null values.
         df = df.with_columns(fuel_cost=pl.col("fuel_consumption") * self.fuel_price)
         self.output["fuel_consumption"].write(df)