+++ src/pymetropolis/metro_demand/od_matrix/custom.py
@@ -1,7 +1,7 @@
 import polars as pl
 
 from pymetropolis.metro_common.io import read_dataframe
-from pymetropolis.metro_demand.modes.car import CarODsFile
+from pymetropolis.metro_demand.routing.files import TripsRoadNodesFile
 from pymetropolis.metro_network.road_network.files import RoadEdgesCleanFile
 from pymetropolis.metro_pipeline.parameters import PathParameter
 from pymetropolis.random import RandomStep
@@ -33,7 +33,7 @@ class CustomODMatrixStep(RandomStep):
         ),
     )
     input_files = {"edges": RoadEdgesCleanFile}
-    output_files = {"car_driver_ods": CarODsFile}
+    output_files = {"road_ods": TripsRoadNodesFile}
 
     def is_defined(self) -> bool:
         return self.file is not None
@@ -42,4 +42,4 @@ class CustomODMatrixStep(RandomStep):
         df = read_dataframe(self.file, columns=["origin", "destination", "size"])
         df = df.filter(pl.col("origin") != pl.col("destination"))
         trips = generate_trips_from_od_matrix(df, self.get_rng())
-        self.output["car_driver_ods"].write(trips)
+        self.output["road_ods"].write(trips)